test: isolate mixed LLDB stack fixtures
